fix(ticketvote): vote status changes cleanup
